202122/Term3/RemoteAccessTool/requirements.txt
pillow
pynput
tqdm